-multiSun commandline flag: Draw sun/skylights only if they come from the sky whose shader they were defined in.
Every shader that emits sun or skyligghts gets an environment emitter index. All the suns/lights generated from that shader get the same index.
trace_t now has a bitmask of all the environment emitter indices encountered when hitting sky surfaces and checking their shader.
Lighting code checks whether the current sunlight has a defined environment emitter index and if it does, it checks whether it is set in the bitmask. Only then will the light be considered to hit.